Status: split heartbeat summary helpers
(cherry picked from commit ebfd32efc31cb26a204358811006dd9d6cc318a9)
